FIX: Adding dry run mode to script, updates to config check and prevent leaving behind replace file alt files. [t34911]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@23505 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/tools/filestore_migrate.php
@@ -2,29 +2,54 @@
 include "../../include/db.php";
 command_line_only();
 
-
-
 // This script moves resources from old locations to new locations if settings are changed after initial setup
 // Useful in in the following situations :-
 //
 // 1) When adding or changing the $scramble_key (setting $migrating_scrambled = true and optionally $scramble_key_old)
 // 2) When enabling $filestore_evenspread 
+//
+// Script will run in dry run mode by default. Supply "commit" as a parameter to the command to apply the changes in the filestore.
+// e.g. filestore_migrate.php commit
 
-if(!$filestore_evenspread && !$migrating_scrambled)
+$config_check = !($filestore_evenspread && $filestore_migrate && $migrating_scrambled) && (($filestore_evenspread && $filestore_migrate) || $migrating_scrambled);
+if (!$config_check)
     {
-    exit("You must manually enable this script by setting \$migrating_scrambled and optionally \$scramble_key_old." . PHP_EOL);
+    exit("You must manually enable this script by setting \$migrating_scrambled and optionally \$scramble_key_old or by setting \$filestore_evenspread and \$filestore_migrate." . PHP_EOL . 
+    "Only one operation is possible at a time. " . PHP_EOL);
     }
-elseif(!$filestore_evenspread || !$filestore_migrate)
+
+if (isset($argv[1]) && $argv[1] == 'commit')
     {
-    exit("\$filestore_evenspread and filestore_migrate must be enabled. Set \$filestore_evenspread=true; and \$filestore_migrate=true; in config.php" . PHP_EOL);     
+    $dry_run = false;
+    echo 'Script running without dry run - the following changes have been applied in the filestore.' . PHP_EOL;
+    }
+else
+    {
+    $dry_run = true;
+    echo 'Script in dry run mode - changes to be made will be shown here but not applied to the filestore.' . PHP_EOL;
+    echo 'To run the script and apply the changes add "commit" to the command e.g. filestore_migrate.php commit' . PHP_EOL;
+    }
+
+if (in_array("rse_version", $plugins)) 
+    {
+    # Don't run script with rse_version enabled - it'll filter out alt files from Replace file so they'll be left behind in the old location.
+    $rse_version_found = array_search("rse_version",$plugins);
+    if ($rse_version_found !== false)
+        {
+        unset($plugins[$rse_version_found]);
+        }
     }
 
 // Flag to set whether we are migrating to even out filestore distibution or because of scramble key change
 $redistribute_mode = $filestore_migrate;
 
-function migrate_files($ref, $alternative, $extension, $sizes, $redistribute_mode)
+# Prevent get_resource_path() attempting to migrate the file as we'll do it in this script when not in dry run mode.
+$filestore_migrate = false;
+$migrating_scrambled = false;
+
+function migrate_files($ref, $alternative, $extension, $sizes, $redistribute_mode, bool $dry_run = true)
     {
-    global $scramble_key, $scramble_key_old, $migratedfiles, $filestore_evenspread, $syncdir;
+    global $scramble_key, $scramble_key_old, $migratedfiles, $filestore_evenspread, $syncdir, $filestore_migrate, $migrating_scrambled;
     echo "Checking Resource ID: " . $ref . ", alternative: " . $alternative . PHP_EOL;
 	$resource_data=get_resource_data($ref);
     $pagecount=get_page_count($resource_data,$alternative);
@@ -55,11 +80,14 @@ function migrate_files($ref, $alternative, $extension, $sizes, $redistribute_mod
                 if(!file_exists($newpath))
                     {
                     echo " - Moving resource file for resource #" . $ref  . " - old path= " . $path  . ", new path=" . $newpath . PHP_EOL;
-                    if(!file_exists(dirname($newpath)))
+                    if (!$dry_run)
                         {
-                        mkdir(dirname($newpath),0777,true);
+                        if(!file_exists(dirname($newpath)))
+                            {
+                            mkdir(dirname($newpath),0777,true);
+                            }
+                        rename ($path,$newpath);
                         }
-                    rename ($path,$newpath);
                     $migratedfiles++;
                     }
                 else
@@ -72,12 +100,10 @@ function migrate_files($ref, $alternative, $extension, $sizes, $redistribute_mod
             if($redistribute_mode)
                 {
                 $filestore_evenspread = true;
-                $filestore_migrate = true;
                 }
             else
                 {
                 $scramble_key = $scramble_key_saved;
-                $migrating_scrambled = true;
                 }
             }
         }
@@ -88,7 +114,10 @@ function migrate_files($ref, $alternative, $extension, $sizes, $redistribute_mod
     if(file_exists($delfolder) && $delfolder != $newfolder && count(scandir($delfolder))==2 && is_writable($delfolder))
         {       
         echo "Deleting folder $delfolder \n";
-        rmdir($delfolder);
+        if (!$dry_run)
+            {
+            rmdir($delfolder);
+            }
         }
 
     }
@@ -112,7 +141,7 @@ for ($n=0;$n<$totalresources;$n++)
     $sizes[] = array("id" => "", "extension" => "xml");
     $sizes[] = array("id" => "", "extension" => "icc");
     
-    migrate_files($ref, -1, $extension, $sizes, $redistribute_mode);
+    migrate_files($ref, -1, $extension, $sizes, $redistribute_mode, $dry_run);
     
     // Migrate the alternatives
     $alternatives = get_alternative_files($ref);
@@ -120,7 +149,7 @@ for ($n=0;$n<$totalresources;$n++)
         {
         $sizes=get_image_sizes($ref,true,$alternative["file_extension"],false);
         $sizes[] = array("id" => "", "extension" => $alternative["file_extension"]);
-        migrate_files($ref, $alternative["ref"], $alternative["file_extension"], $sizes, $redistribute_mode);
+        migrate_files($ref, $alternative["ref"], $alternative["file_extension"], $sizes, $redistribute_mode, $dry_run);
         }
     }
